+++ z0lib/z0lib/scripts/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import main
+from .main import get_metadata